refactor: moving the networking specific method out
